+++ lib/zstd_errors.h
